% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how.old.R
\name{how.old}
\alias{how.old}
\title{Determine numeric age}
\usage{
how.old(from, to)
}
\arguments{
\item{from}{character, beginning of the date range in \code{YYYY-MM-DD} format.}

\item{to}{character, end of the date range in \code{YYYY-MM-DD} format.}
}
\value{
Numeric age in years.
}
\description{
Determines the numeric age in years from a date range.
}
\details{
Helper for LexisPlotR. The numeric age gets rounded to 5 digits.
}
\examples{
library(LexisPlotR)
how.old(from = "1900-01-01", to = "1905-01-01")
}
\author{
Philipp Ottolinger
}

