% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexis.survey.R
\name{lexis.survey}
\alias{lexis.survey}
\title{Deprecated. Emphasize a survey range in a Lexis grid
Takes an existing Lexis grid and adds a coloured parallelogram to highlight a survey range.}
\usage{
lexis.survey(lg, from_date, to_date, from_age, to_age,
  fill = lpr_colours()[6], alpha = 0.7)
}
\arguments{
\item{lg, }{an existing object originally created with \code{lexis.grid()}.}

\item{from_date}{character, set the beginning of the survey in format \code{"YYYY-MM-DD"}.}

\item{to_date}{character, set the end of the survey in format \code{"YYYY-MM-DD"}.}

\item{from_age}{numeric, set the starting age of the survey.}

\item{to_age}{numeric, set the ending age of the survey.}

\item{fill}{character, set the colour to fill the parallelogram. Default is \code{"orange"}.}

\item{alpha}{numeric, set the transparency of the fill colour. Default is \code{0.5}.}
}
\description{
Deprecated. Emphasize a survey range in a Lexis grid
Takes an existing Lexis grid and adds a coloured parallelogram to highlight a survey range.
}
\details{
The function can be used to plot the time and age range of a survey. Use \code{from_date} and \code{to_date} to specify the time range the survey took place and \code{from_age} and \code{to_age} to define the considered ages of participants/observations.
}
\examples{
\dontrun{
library(LexisPlotR)
lg <- lexis.grid(year.start = 1980, year.end = 1990, age.start = 30, age.end = 40)
lexis.survey(lg, from_date = "1982-09-01", to_date = "1986-03-01", from_age = 32, to_age = 36)
}
}
\author{
Philipp Ottolinger
}
