\name{LexisPlotR-package}
\alias{LexisPlotR-package}
\alias{LexisPlotR}
\docType{package}
\title{
\packageTitle{LexisPlotR}
}
\description{
\packageDescription{LexisPlotR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LexisPlotR}
\packageIndices{LexisPlotR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LexisPlotR}

Maintainer: \packageMaintainer{LexisPlotR}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
