% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lifelines_sample}
\alias{lifelines_sample}
\title{Data for 300 random lifelines}
\format{A data frame with 300 rows and 2 variables:
\describe{
  \item{entry}{entry or birth dates.}
  \item{exit}{exit or death dates, \code{NA} if not observed.}
}}
\usage{
lifelines_sample
}
\description{
This dataset contains 300 random entry dates and 150 exit dates for demonstration purposes.
}
\keyword{datasets}

