% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what.date.R
\name{what.date}
\alias{what.date}
\title{Determine the date a certain age is reached}
\usage{
what.date(date, age)
}
\arguments{
\item{date}{character, set the reference date in \code{YYYY-MM-DD} format.}

\item{age}{numeric, set an age to be reached.}
}
\value{
The date \code{age} is reached when counting from \code{date}.
}
\description{
Determines the date a certain age is reached given an earlier date.
}
\details{
Helper for LexisPlotR.
}
\examples{
library(LexisPlotR)
what.date(date = "1900-01-01", age = 3)
}
\author{
Philipp Ottolinger
}

