% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LexisPlotR-package.R
\docType{package}
\name{LexisPlotR-package}
\alias{LexisPlotR}
\alias{LexisPlotR-package}
\title{c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_title(\"#1\")}",
"LexisPlotR")\Sexpr{tools:::Rd_package_title("LexisPlotR")}}
\description{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_description(\"#1\")}",
"LexisPlotR")\Sexpr{tools:::Rd_package_description("LexisPlotR")}
}
\details{
The DESCRIPTION file:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_DESCRIPTION(\"#1\")}",
"LexisPlotR")\Sexpr{tools:::Rd_package_DESCRIPTION("LexisPlotR")}
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_indices(\"#1\")}",
"LexisPlotR")\Sexpr{tools:::Rd_package_indices("LexisPlotR")} ~~ An overview
of how to use the package, including the most important functions ~~
}
\examples{


}
\author{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_author(\"#1\")}",
"LexisPlotR")\Sexpr{tools:::Rd_package_author("LexisPlotR")}

Maintainer:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_maintainer(\"#1\")}",
"LexisPlotR")\Sexpr{tools:::Rd_package_maintainer("LexisPlotR")}
}
\keyword{package}

